Make /reload update reported branch name too
